fix: move formatting utils to src/utils/ to fix reversed imports (#225)
services/ and utils/ were importing markdown_to_telegram_html and
strip_telegram_html from src/bot/handlers/formatting, violating the
layering rule. Canonical code now lives in src/utils/formatting.py;
the bot handler module re-exports for backward compat.

Adds test_import_layering.py to enforce the rule going forward.
